Ensure all contained associations are plugin-ready
When CakePHP initialises the tables in this plugin, they all use the
default datasource. This is a problem when Opencart database requires
a different datasource.

This will for each table initialised inside this plugin:

- set the right connection
- set the right entity class
- set the default language is this is a description association

("Description" means selecting the translation in the default language
as opposed to "Descriptions" selecting all translations.)