Fix singlular description association table name
It used to link to singular table names but it stopped working after Connector code was moved to cms-connector plugin. The fix seems to be to tell the tables that singular description should use plural table names (which actually makes sense, as singular tables e.g. ProductDescription do not exist anyway).